Update Subprocess Functions and Add Sort/Index Function
The subprocess functions were updated to use OSError rather than CalledProcessError. The latter contains too much detail about the failed process than is needed

A function for sorting and indexing the aligned file was added. It currently doesn't work as the input needs to be converted to BAM format first